Move the dummy data to another file
