+++ frontend/src/pages/QuizPage.js
@@ -6,6 +6,7 @@ import QuizButton from "../components/quizButton";
 import Checkbox from "../components/checkbox";
 import SubmitButton from "../components/submitButton";
 import iconMap from "../lib/iconMap";
+import quiz from "../lib/quiz";
 
 export default function QuizPage() {
   const { t } = useTranslation("quiz");
@@ -15,98 +16,6 @@ export default function QuizPage() {
   const [allAnswer, setAllAnswer] = useState([]);
 
   const [step, setStep] = useState(0);
-
-  // hardcoded value to mockup the data we will receive from the backend
-  // TO REPLACE BY API REQUEST
-  const quiz = [
-    {
-      scope: { id: 1, filter: "user" },
-      answers: [
-        { id: 1, text: "you", value: "you" },
-        { id: 2, text: "else", value: "else" },
-      ],
-      type: "radio",
-    },
-    {
-      scope: { id: 2, filter: "age" },
-      answers: [
-        { id: 1, text: "baby", value: "baby" },
-        { id: 2, text: "child", value: "child" },
-        { id: 3, text: "youth", value: "youth" },
-        { id: 4, text: "adult", value: "adult" },
-        { id: 5, text: "skip", value: "" },
-      ],
-      type: "radio",
-    },
-    {
-      scope: { id: 3, filter: "gender" },
-      answers: [
-        { id: 1, text: "male", value: "male" },
-        { id: 2, text: "female", value: "female" },
-        { id: 3, text: "other", value: "other" },
-        { id: 4, text: "skip", value: "" },
-      ],
-      type: "radio",
-    },
-    {
-      scope: { id: 4, filter: "region" },
-      answers: [
-        { id: 1, text: "dakar", value: "dakar" },
-        { id: 2, text: "diourbel", value: "diourbel" },
-        { id: 3, text: "kaolack", value: "kaolack" },
-        { id: 4, text: "saintlouis", value: "saintlouis" },
-        { id: 5, text: "thies", value: "thies" },
-        { id: 6, text: "all", value: "all" },
-      ],
-      type: "checkbox",
-    },
-    {
-      scope: { id: 5, filter: "dots_categories" },
-      answers: [
-        {
-          id: 1,
-          text: "socialProtection",
-          value: "1",
-        },
-        {
-          id: 2,
-          text: "food",
-          value: "2",
-        },
-        {
-          id: 3,
-          text: "productivityAndIncome",
-          value: "3",
-        },
-        {
-          id: 4,
-          text: "wellBeing",
-          value: "4",
-        },
-        {
-          id: 5,
-          text: "education",
-          value: "5",
-        },
-      ],
-      type: "checkbox",
-    },
-    {
-      scope: { id: 6, filter: "dots_subCategories" },
-      answers: [
-        { id: 1, text: "housingAndSocialServices", value: "socialProtectionOne" },
-        { id: 2, text: "economicResources", value: "socialProtectionTwo" },
-        {
-          id: 3,
-          text: "financialServices",
-          value: "socialProtectionThree",
-        },
-        { id: 4, text: "protectionClimatEvents", value: "socialProtectionFour" },
-        { id: 5, text: "alleviatePoverty", value: "socialProtectionFive" },
-      ],
-      type: "checkbox",
-    },
-  ];
 
   // when user can have more than one answer, makes an array with multiple answers
   // which will be deleted when they untick the answer
